+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ability to pass in a context argument to steps (#2)
+
 ## [0.5.0] - 2023-01-13
 
 ### Added